move swiper arrows to folder /shared/assets
