+++ frontend/src/client/widgets/employerBlock/ui/EmployerBlock.tsx
@@ -3,8 +3,8 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation } from 'swiper/modules';
 import { Box, Typography } from '@mui/material';
 import EmployerCard from '../../../feachers/employerCard/ui/EmployerCard';
-import arrowLeft from '../../MediaBlock/images/arrow-left.png';
-import arrowRight from '../../MediaBlock/images/arrow-right.png';
+import arrowLeft from '../../../../shared/assets/arrow-left.png';
+import arrowRight from '../../../../shared/assets/arrow-right.png';
 import EmployerBlockStyle from './EmployerBlock-style';
 import './SwiperNavigationEmployerBlock.css';
 import 'swiper/css';
