+++ frontend/src/shared/assets/arrow-left.png
