+++ frontend/src/shared/assets/arrow-right.png
